Made the timer into a separate class
