+++ Parking/Parking/Parking.csproj
@@ -49,6 +49,7 @@
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Settings.cs" />
+    <Compile Include="Timer.cs" />
     <Compile Include="Transaction.cs" />
   </ItemGroup>
   <ItemGroup>
